// models/User.ts
import { BelongsToManyGetAssociationsMixin, CreationOptional, InferAttributes, InferCreationAttributes, Model, NonAttribute } from 'sequelize';
import { sequelize, DataTypes } from '../utils/database';
import { RoomModel } from './Room';
import { MessageModel } from './Message';

export interface UserModel extends Model<InferAttributes<UserModel>, InferCreationAttributes<UserModel>> {
    id: CreationOptional<string>,
    userName: string,
    email: string,
    avatar: CreationOptional<string>,
    hash: CreationOptional<string>,
    status: CreationOptional<string>
    createdAt: CreationOptional<Date>,
    updatedAt: CreationOptional<Date>,

    // mixin methods (these are auto-generated by sequelize we only define signatures for TS)
    getRooms: BelongsToManyGetAssociationsMixin<RoomModel>,
    Rooms?: RoomModel[],

    Messages?: MessageModel[]
}

const User = sequelize.define<UserModel>('User', {
    id: {
        primaryKey: true,
        type: DataTypes.UUID,
        defaultValue: DataTypes.UUIDV4,
        allowNull: false
    },
    email: {
        type: DataTypes.STRING,
        allowNull: false,
        unique: true
    },
    userName: {
        type: DataTypes.STRING,
        allowNull: false
    },
    avatar: {
        type: DataTypes.STRING,
        defaultValue: "",
        allowNull: false
    },
    hash: {
        type: DataTypes.STRING,
        allowNull: false
    },
    status: {
        type: DataTypes.STRING,
        defaultValue: "Hello there!",
        allowNull: false,
    },
    createdAt: {
        type: DataTypes.DATE
    },
    updatedAt: {
        type: DataTypes.DATE
    },
}, {
    tableName: "users"
});

export default User;